Replaced AVAudioPlayer with AVPlayer
Major change to implementation of AudioPlayer. AudioPlayer no longer uses an instance of an AVAudioPlayer. Instead, it now inherits from AVPlayer. An instance of AudioPlayer is injected into the environment so it is accessible by all views presented by ContentView. Published property wrapper used for current time in player to ensure all views using this property are updated appropriately